separate getContentType and return a promise
